cleanup: Merge crypto_core and crypto_core_mem.
The remaining memory functions are small and don't need their own file.
