+++ other/bootstrap_daemon/docker/tox-bootstrapd.sha256
@@ -1,1 +1,1 @@
-fc6b060abddd1898d97b0cf067d19c7a9d68999469690d91e7cf59327e700dbf  /usr/local/bin/tox-bootstrapd
+243145e05d7d20b7703a339ae0898d0445400ef2ecc852096f13e32b1609912c  /usr/local/bin/tox-bootstrapd
